Move machine into its own file
